+++ R/import.R
@@ -1,16 +1,14 @@
 #' @import nlreferences
 #' @importFrom bdsreader      persondata timedata
 #' @importFrom centile        y2z
-#' @importFrom dplyr          %>% .data arrange bind_rows case_when filter
-#'                            full_join if_else left_join mutate mutate_at recode
-#'                            rename select transmute slice vars
-#' @importFrom dplyr          %>% .data arrange bind_rows distinct filter full_join
-#'                            left_join mutate mutate_at recode rename select
-#'                            transmute slice summarise vars
+#' @importFrom dplyr          %>% .data arrange bind_rows case_when distinct
+#'                            filter full_join if_else left_join mutate
+#'                            mutate_at recode rename select transmute slice
+#'                            summarise vars
 #' @importFrom methods        slot
 #' @importFrom rlang          .data
 #' @importFrom stats          approx
 #' @importFrom tidyselect     all_of starts_with
-#' @importFrom tidyr          fill tibble
+#' @importFrom tidyr          fill pivot_wider tibble
 #' @importFrom utils          hasName packageVersion
 NULL