+++ NEWS.md
@@ -1,3 +1,6 @@
+
+- Improved `calculate_advice_devlang` code
+
 # growthscreener 1.24.0
 
 - Fixes issue in hgt advice and correct date in examples (#5)